Add locale override setting with validation
